pyproject.toml
[project]
name = "beeai-requirement-agent-examples"
version = "0.1.0"
description = "Practical examples for the BeeAI Requirement Agent"
readme = "README.md"
authors = [
    { name = "BeeAI Community" }
]
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
dependencies = [
    "beeai-framework>=0.1.36",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["examples"]

[tool.ruff]
line-length = 120
target-version = "py311"